+++ R/points_in_circle.R
@@ -1,6 +1,6 @@
 #' Points in circle
 #'
-#' @description All observations within a circle of a certain radius.
+#' @description Find all observations in a data.frame within a circle of a certain radius.
 #'
 #' @param data data.frame with at least columns for longitude and latitude.
 #' @param lon_center numeric value referencing to the longitude of the center of the circle
@@ -19,7 +19,7 @@
 #'
 #' @return data.frame. Column \code{distance_m} gives the distance to the center of the circle (in meters).
 #'
-#' @export points_in_circle
+#' @export
 points_in_circle <- function(data, lon_center, lat_center, lon = lon, lat = lat, radius = 200){
 
   # Turn into character vector
@@ -28,11 +28,11 @@ points_in_circle <- function(data, lon_center, lat_center, lon = lon, lat = lat,
   data_name <- deparse(substitute(data))
 
   if ( !all(c(lon, lat) %in% names(data))) {
-    stop(paste0(data_name, " does not contain columns ", lon, " and ", lat))
+    stop(paste0(data_name, " does not contain columns ", lon, " and ", lat), call. = FALSE)
   }
 
   if ( !all(is.numeric(c(data[[lon]], data[[lat]]))) ){
-    stop(paste0(lon, ", ", lat, " should be numeric"))
+    stop(paste0(lon, ", ", lat, " should be numeric"), call. = FALSE)
   }
 
   df <- data.frame("lon" = data[[lon]], "lat" = data[[lat]])