+++ NEWS.md
@@ -1,3 +1,8 @@
+# spatialrisk 0.6.8
+
+* `highest_concentration()` is added to do a fast search for the coordinates of the highest concentration
+* `neighborhood_gh_search()` is added to look for even higher concentrations in the neighborhood of the coordinates found by `highest_concentration()`
+
 # spatialrisk 0.6.7
 
 * A package website is added using pkgdown.